채팅방리스트와 채팅방 분리
