+++ app/chat/chat.module.scss
@@ -24,11 +24,14 @@
     border-radius: 4px;
 
     & .chat_room_item {
-      width: 100%;
-      border-bottom: 1px solid #ddd;
-      display: flex;
-      padding: 24px 16px;
-      align-items: center;
+      & a {
+        width: 100%;
+        border-bottom: 1px solid #ddd;
+        display: flex;
+        padding: 24px 16px;
+        align-items: center;
+        background: #fff;
+      }
 
       & .profile_image {
         width: 50px;
@@ -52,6 +55,7 @@
         margin-right: 16px;
         gap: 6px;
         flex: 1;
+        align-items: start;
 
         & .title {
           font-size: 16px;
@@ -63,133 +67,6 @@
           font-weight: 400;
         }
       }
-    }
-  }
-
-  & .chat_section {
-    @include mixin.flexCenter(column);
-    gap: 40px;
-    max-width: 500px;
-
-    & .section_title {
-      display: flex;
-      align-items: center;
-      gap: 24px;
-    }
-  }
-
-  & .item_container {
-    display: flex;
-    flex-direction: column;
-    gap: 24px;
-    max-height: 500px;
-    overflow-y: scroll;
-    min-width: 500px;
-
-    & .chat_item {
-      display: flex;
-      flex-direction: column;
-      background: rgba(102, 126, 234, 0.1);
-      padding: 16px;
-      border-radius: 12px;
-
-      & .content_top {
-        display: flex;
-        justify-content: space-between;
-        margin-bottom: 6px;
-
-        & .chat_title {
-          @include mixin.flexCenter(row);
-          gap: 8px;
-
-          & .profile_image {
-            width: 24px;
-            height: 24px;
-            border-radius: 50%;
-            background: burlywood;
-          }
-
-          & .chat_name {
-            font-weight: 600;
-            font-size: 14px;
-          }
-
-          & .chat_bedge {
-            padding: 3px 8px;
-            background: #667eea;
-            color: #fff;
-            display: inline-block;
-            @include mixin.flexCenter(row);
-            border-radius: 4px;
-            font-size: 12px;
-          }
-        }
-
-        & .date {
-          color: #555;
-          font-size: 12px;
-        }
-      }
-
-      & .content_bottom {
-        display: flex;
-        flex-direction: column;
-        gap: 8px;
-        border: 1px solid #ddd;
-        padding: 8px;
-        border-radius: 8px;
-        align-items: start;
-
-        & .chat_type {
-          @include mixin.flexCenter(row);
-          padding: 2px 4px;
-          background: #667eea;
-          color: #fff;
-          border-radius: 4px;
-          font-size: 12px;
-        }
-
-        & .chat_content {
-          padding: 8px;
-          border-radius: 8px;
-          font-size: 14px;
-        }
-
-        & .content_bottom_title {
-          display: flex;
-          gap: 12px;
-
-          & button {
-            @include mixin.flexCenter(row);
-            padding: 2px 4px;
-            color: #667eea;
-            border: 1px solid #667eea;
-            border-radius: 4px;
-            font-size: 12px;
-            background: #fff;
-          }
-        }
-      }
-    }
-  }
-
-  & .chat_enter {
-    display: flex;
-    width: 100%;
-    position: relative;
-    gap: 8px;
-    flex-direction: column;
-    align-items: end;
-
-    & .submit_button {
-      width: 40px;
-      height: 33px;
-      background: #ededed;
-      color: #333;
-    }
-
-    & .text_field {
-      border-radius: 4px;
     }
   }
 }
