+++ app/mentor/page.tsx
@@ -6,9 +6,14 @@ import { useEffect, useState } from 'react';
 import { mentorService } from '../services/mypage/mentor';
 import { Mentors } from '../types/mentor/search';
 import { chatroomService } from '../services/chatroom/chatroom';
+import { RootState } from '../store/store';
+import { useSelector } from 'react-redux';
+import { Member } from '../store/isLogin/loginSlice';
 
 const MentorPage = () => {
+  const member = useSelector((state: RootState) => state.login.member);
   const [mentors, setMentors] = useState<Mentors[]>([]);
+  const [user, setUser] = useState<Member>(member);
 
   const { getMentorList } = mentorService();
   const { addChatRoom } = chatroomService;
@@ -19,13 +24,17 @@ const MentorPage = () => {
       description: '멘티가 신청한 채팅방 내용123123213636436',
       created_member_id: createdId,
       max_people: 3,
-      member_id: 'c4b1ac92-d06d-4fa7-b2c1-1738b2b4cab4',
+      member_id: user?.id,
     };
 
     addChatRoom(chatRoomData).then((res) => {
       console.log('res', res);
     });
   };
+
+  useEffect(() => {
+    setUser(member);
+  }, [member]);
 
   useEffect(() => {
     getMentorList().then((res) => {
